HomeWork13 ДЗ на проверку добавлены toString, equals и hashCode
